% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks}
\alias{breaks}
\title{Function to find best number of intervals iteratively. Works with cut methods 
c("numbers", "width")}
\usage{
breaks(x = x, equal = "number", nbins = 6, binwidth = NULL)
}
\description{
Function to find best number of intervals iteratively. Works with cut methods 
c("numbers", "width")
}
