% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_S_list.R
\name{build_S_list}
\alias{build_S_list}
\title{Build list for cross-talk calculation}
\usage{
build_S_list(S_tab, g_u, universe, min_size = 1, max_size = 500)
}
\arguments{
\item{S_tab}{a table composed by two (mandatory) or three columns: the first contain the names of the gene sets, the second the name of the genes. The third column
(if provided) should be the gene weights. If `ncol(S_tab) == 3`, `g_u` param is ignored
weights in the gene sets (i.e.: cell-cell communication), there should be a third column with the weights of each gene in the gene sets.}

\item{g_u}{named vector. Should contain the genes (as names) and their weights. Ignored if `ncol(S_tab) == 3`.}

\item{universe}{genes in the adjacency matrix that will be used in `gs_cross_talk()`. Set to `NULL` to skip filtering}

\item{min_size, max_size}{filtering the gene sets to maintain only the ones that 
have at least `min_size` or less than `max_size` number of genes. Set to `NULL` to skip filtering}
}
\value{
The output is a gene-set list with a vector for each gene-set, composed by weights named after the genes
}
\description{
The function build a gene-set list compatible with `gs_cross-talk()` for cross-talk calculation starting from a provided gene set database
 and genes (with weights) of interest.
}
\details{
`build_S_list()` uses a gene-set database data to build the list needed by `gs_cross_talk()` for cross-talk calculation.
}
