% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gs_perm.R
\name{calc_gs_perm}
\alias{calc_gs_perm}
\title{Internal function of pathtools}
\usage{
calc_gs_perm(rll = NULL, perm = NULL, gs = NULL)
}
\arguments{
\item{rll}{list of named ranked vectors}

\item{perm}{vector of permuted names}

\item{gs}{gene set}
}
\description{
Internal function of pathtools
}
