% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p.R
\name{calc_p}
\alias{calc_p}
\title{Estimation of p values}
\usage{
calc_p(X)
}
\arguments{
\item{X}{list of matrices, where the first is the one obtained with real data and the others are obtained through permutations. 
The matrices should be 1 column matrices with the same order.}
}
\description{
This function is used to calculate p-value from a list of matrices
}
\details{
the functions calculate p-value for a permutation-based approach. It takes as an input a list of matrices, where the first is the one with the real data
and the other are the data obtained through permutations. Then, the function searches for how many values in the permutations are equal or higher to the real.
These values are then divided by the number matrices (1 + number of permutations)
}
