% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_set_similarity.R
\name{calc_set_similarity}
\alias{calc_set_similarity}
\title{Similarity between two sets}
\usage{
calc_set_similarity(x, y, method = c("jaccard", "overlap", "average"))
}
\arguments{
\item{x, y}{two sets}

\item{method}{one among "jaccard", "overlap" and "average"}
}
\value{
similarity
}
\description{
Similarity between two sets
}
