% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_communication.R
\name{cluster_communication}
\alias{cluster_communication}
\title{Caluclate communication between two clusters}
\usage{
cluster_communication(
  DEG_list,
  gene_network_adj,
  k = 9,
  ligand = ligand,
  receptor = receptor,
  mc_cores_perm = 1
)
}
\arguments{
\item{DEG_list}{a gene list composed by a vector for both clusters. The vector should be a weight vector named by gene}

\item{gene_network_adj}{gene network adjacency matrix}

\item{k}{number of permutation of the adjacency matrix}

\item{ligand}{vector of ligand genes}

\item{receptor}{vector of receptor genes}

\item{mc_cores_perm}{number of threads to be used to calcualte permutation}
}
\description{
Caluclate communication between two clusters
}
