% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_results_network.R
\name{comparing_results_network}
\alias{comparing_results_network}
\title{Function to plot the union network of multiple PCT or CCC analyses by highlighting the shared edges and vertices.}
\usage{
comparing_results_network(
  res_list,
  file_out,
  colors = NULL,
  value = NULL,
  edge_width = T,
  legend_col = T,
  legend_edge_width = T,
  layout = NULL,
  width = 200,
  height = 200,
  res = 300,
  units = "mm",
  x_vertex = -1.5,
  y_vertex = 0,
  x_edge_w = -1.5,
  y_edge_w = -0.3,
  xl = -1.45,
  yb = -0.9,
  xr = -1.3,
  yt = -0.6,
  ...
)
}
\arguments{
\item{res_list}{list of resulting data.frame from `pathway_cross_talk()` or `cluster_communication()`}

\item{file_out}{filename of the produced plot. If not provided the plot is saved as "union_ct_network.jpeg"}

\item{colors}{colors to be used for the vertices. If `vertex_col_by = number` should be a vector of two colors. If `vertex_col_by` is a vector,
then should be a vector of colors, one for each unique variable in the `vertex_col_by` vector. If set to `NULL` then the function uses blue to red for `number`
and rainbow palette for vector}

\item{value}{value to be used for the gradient of the vertices when `vertex_col_by = number`. If `NULL` the function uses min and max}

\item{edge_width}{logical, if the edge width should change according to how much it is shared in the results list.}

\item{legend_col}{logical, if to plot the legend of the color of the edges and vertices}

\item{legend_edge_width}{logical, if to plot the legend of the different interval used for edge width plotting}

\item{layout}{layout to be used in igraph plotting. If not provided, the function calculates it}

\item{width, height, res, units}{graphical value of `jpeg()` function}

\item{x_vertex, y_vertex, x_edge_w, y_edge_w}{x and y coordinates for `legend_vertex` (when a vector is passed to `vertex_col_by` ) and `legend_edge_breaks` 
positioning in the plot. See `legend()` function for more details}

\item{xl, yb, xr, yt}{coordinates for the positioning of the gradient legend for `legend_col`. See `color.legend()` 
and `rect()` functions for more details}

\item{...}{further argument passed to `igraph.plot` function}

\item{legend_vertex}{logical, if to print the legend of the color on the vertices}
}
\value{
the function produce the plot saved with the name passed to `file_out` and also returns the layout, and the igraph object used for plotting
}
\description{
The function is used to visually compare the results obtained by multiple PCR or CCC analyses by creating a union network and highlighting the
shared edges
}
\details{
The function takes as an input a list of data.frame resulting from multiple `pathway_cross_talk()` or `cluster_communication()` analyses
and produce a union network. This network is plotted by coloring the edges corresponding to how many analyses are present. The user may also decide to color
the vertices on the number of samples in which are present or by a user-provided variable
}
