% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFDR.R
\name{eFDR}
\alias{eFDR}
\title{empirical False Discovery Rate}
\usage{
eFDR(real_values, all_values, mc.cores = 1)
}
\arguments{
\item{real_values}{vactor of real (observed) values}

\item{all_values}{vactor of all values (real + permuted) values}

\item{mc.cores}{number of cores}
}
\description{
empirical False Discovery Rate
}
