% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_communities.R
\name{find_communities}
\alias{find_communities}
\title{Find topological communities}
\usage{
find_communities(
  g,
  e.weights = NULL,
  v.weights = NULL,
  verbose = TRUE,
  methods = c("fastgreedy", "labprop", "walktrap", "eigen", "multilev", "infomap")
)
}
\arguments{
\item{g}{graph}

\item{e.weights}{edge weights}

\item{v.weights}{vertex weights}

\item{verbose}{TRUE/FALSE}

\item{methods}{one or more of "fastgreedy", "labprop", "walktrap", "eigen", "multilev", "infomap"}
}
\value{
list od community objects
}
\description{
Find topological communities
}
