% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_classification.R
\name{gene_classification}
\alias{gene_classification}
\title{Cross-talk gene classification analysis}
\usage{
gene_classification(gs_ct = NULL, A = NULL, ct_ref = NULL, n_cores = 1)
}
\arguments{
\item{gs_ct}{the data.frame obtained as a result of `gs_cross_talk()`, filtered if needed}

\item{A}{the adjacency matrix used as an input in `gs_cross_talk()`}

\item{ct_ref}{=`NULL` or a data.frame with cross-talk calculated on a reference. 
If provided, saturation `r_g` and `r_d` are calculated.}

\item{n_cores}{number of cores to be use to parallelize gene gene set relevance analysis}
}
\value{
If `ct_ref = NULL` The function returns a data.frame with:
\itemize{
\item gene: the gene analysed
\item d_q: number of altered cross-talk `gene` contributes to
\item d_d: number of interactors contributing to any altered cross-talk together with `gene`
\item d_q_S: list of names of the gene-sets counted in `d_q` separated by `;`
\item d_d_gene: list of names of the genes counted in `d_d` separated by `;`
} 
Otherwise, if `ct_ref` is provided the function returns a data.frame with:
\itemize{
\item gene: the gene analysed
\item d_q: cross-talk diversity, or number of gene set in altered cross-talk that contains interactors of `gene`
\item d_d: interaction diversity, or number of interactors of `gene` that are part of altered cross-talks
\item r_q: saturation of cross-talk diversity
\item r_d: saturation of interaction diversity
\item d_q_S: list of gene set names counted in `d_q` separated by `;`
\item d_d_gene: list of names of the genes counted in `d_d` separated by `;`
\item q: number of gene set in cross-talk that contains interactors of `gene`
\item d: number of interactors of `gene` that are part of cross-talks
\item q_S: list of gene set names counted in `q` separated by `;`
\item d_gene: list of names of the genes counted in `d` separated by `;`
}
}
\description{
The function calculates (saturation of) cross-talk diversity and (saturation of) 
 interaction diversity involved in provided cross-talks
}
\details{
The function takes as an input the data.frame obtained from `gs_cross_talk()` and the 
 adjacency matrix used as an input of `gs_cross_talk()`. These inputs are used to calculate 
 (saturation of) cross-talk diversity and (saturation of) interaction diversity
 involved in significant cross-talks, as defined in the paper. 
 This function has two applications:
 \enumerate{
 \item calculate cross-talk diversity and interaction diversity in a reference model
 \item calculate saturation of cross-talk diversity and saturation of interaction diversity 
   by using the reference obtained in `1`
 }
}
