% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TCGA.R
\name{get_TCGA}
\alias{get_TCGA}
\title{Obtaining data from TCGA for vignette}
\usage{
get_TCGA(
  project,
  data.category,
  data.type,
  sample.type,
  workflow.type,
  summarizedExperiment,
  save
)
}
\arguments{
\item{project}{name of the TCGA project}

\item{data.category}{category of the data}

\item{data.type}{type of data}

\item{sample.type}{type of sample}

\item{workflow.type}{type of workflow}

\item{summarizedExperiment}{logical indicating if the data should be imported as a summarized experiment}

\item{save}{logical, save or not the downloaded data locally}
}
\description{
Obtaining data from TCGA for vignette
}
