% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea2enrich.R
\name{gsea2enrich}
\alias{gsea2enrich}
\title{Create a gseaResult instance from Ulisse GSEA result}
\usage{
gsea2enrich(
  gsea_res = NULL,
  rl = NULL,
  gsl = NULL,
  min_size = 5,
  max_size = 500,
  decreasing = NULL
)
}
\arguments{
\item{gsea_res}{output of function gsea()}

\item{rl}{the input used to obtain gsea_res; numeric matrix of genes-by-ranking criteria; each column contains numeric values; rownames are mandatory;}

\item{gsl}{named list of gene sets used to obtain gsea_res}

\item{min_size}{minimum gene set size}

\item{max_size}{maximum gene set size}

\item{decreasing}{TRUE/FALSE vector that specifies whether to order each column of rl decreasingly or not; must be of length equal to `ncol(rl)`. If NULL, all #' @export}
}
\description{
Create a gseaResult instance from Ulisse GSEA result
}
