% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_cross_talk.R
\name{integrated_cross_talk}
\alias{integrated_cross_talk}
\title{Integrated cellular cross talk analysis}
\usage{
integrated_cross_talk(
  cl_list = cl_list,
  ccc_list = ccc_list,
  gs_list = ptw_list,
  A = adj.ppi,
  k = 9,
  perm_link = "degree",
  perm_weights = "degree",
  cut_par = 3,
  mc_cores_perm = 1,
  mc_cores_ct = 1
)
}
\arguments{
\item{cl_list}{cluster list used for communication analysis (provided as `S_list` to `gs_cross_talk()`)}

\item{ccc_list}{communication details between cluster of interest}

\item{gs_list}{list of gene set lists, one for each cluster considered in the analysis}

\item{A}{adjacency matrix of the whole gene network considered (can be a sparseMatrix)}

\item{k}{number of permutations}

\item{perm_link}{= c("degree", "simple") If the permutation of the link should be degree-conservative ("degree"), or random ("simple")}

\item{perm_weights}{= c("degree", "simple") If the permutation of the weights should be degree-conservative ("degree"), or random ("simple")}

\item{cut_par}{number of bin to cut. If set to `NULL`, the function will search for the best cut in 2:15}

\item{mc_cores_perm}{number of thread to be used in permutations}

\item{mc_cores_ct}{number of threads to be used for cross talk calculation}

\item{bin_type}{can be either "number" (suggested), "interval" or "width". See `ggplot2::cut_interval()` for details.}
}
\value{
The function returns a data.frame with the result of inter-intracellular communication:
\itemize{
   \item ccc: cluster pair considered for integrated cross-talk analysis
   \item S1_name,S2_name: names of gene-set pairs considered
   \item c: cross-talk score
   \item S1_size,S2_size: number of genes in S1 and S2
   \item S1_s2_size,S2_s1_size: number of S1 genes interacting with S2, and vice versa
   \item dL: number of links between S1 and S2
   \item L: number of possible links between S1 and S2
   \item r_c: cross-talk saturation, calculated as `dL/L`
   \item u1,u2: sum of the gene weights in S1 and S2, respectively
   \itme S1,S2: list of interacting genes in S1 and S2, respectively
   \item s: cross-talk summary score
   \item pA,pU: p-values of the number of links (pA) and weights (pU)
   \item p: combined p-value
 }
}
\description{
The function is a wrapper of `single_integrated_cross_talk()` and calculates inter-intracellular communication on multiple clusters.
}
\details{
The function takes as inputs takes as input the list of communication details of interest (`ccc_list`), 
 together with the cluster-list used for CCC calculation (`cl_list`), a list of gene sets lists (`gs_list`), one for each cluster/cell type,
 and the adjacency matrix of the biological network (`A`). 
 These inputs are formatted and passed to `single_integrated_cross_talk()` to calculate integrated cross-talk on each cluster pair provided.
}
