% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicl_communication.R
\name{multicl_communication}
\alias{multicl_communication}
\title{Caluclate communication between multiple clusters}
\usage{
multicl_communication(
  DEG_list,
  gene_network_adj,
  k = 9,
  ligand = ligand,
  receptor = receptor,
  mc_cores_perm = 1
)
}
\arguments{
\item{DEG_list}{a list composed by gene list for each couple of clusters. Each gene list should be composed by weight vector named by gene for each cluster}

\item{gene_network_adj}{gene network adjacency matrix}

\item{k}{number of permutation of the adjacency matrix}

\item{ligand}{vector of ligand genes}

\item{receptor}{vector of receptor genes}

\item{mc_cores_perm}{number of threads to be used to calcualte permutation}
}
\description{
Caluclate communication between multiple clusters
}
