% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{ora}
\alias{ora}
\title{Over Representation Analysis}
\usage{
ora(
  wb = NULL,
  universe = NULL,
  gsl = NULL,
  p_adj_method = "fdr",
  description = NULL,
  wbd_min = 1,
  min_size = 5,
  max_size = 500,
  out_file_prefix = "ora_res"
)
}
\arguments{
\item{wb}{list of character vectors with hits (white balls)}

\item{universe}{universe, character vector}

\item{gsl}{named list of sets}

\item{p_adj_method}{p value adjustment method, see p.adjust.methods}

\item{description}{optional named vector with gene set description; names must be gene set identifiers}

\item{wbd_min}{minimum number of white balls drawn}

\item{min_size}{minimum gene set size}

\item{max_size}{maximum gene set size}

\item{out_file_prefix}{prefix for .xlsx and .txt output files}
}
\description{
Over Representation Analysis
}
