% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora2enrich.R
\name{ora2enrich}
\alias{ora2enrich}
\title{ora2enrich
this function translates the result of an ORA into the DOSE class "enrichResult"}
\usage{
ora2enrich(
  ora_res = NULL,
  pvalueCutoff = 1,
  pAdjustMethod = "BH",
  qvalueCutoff = 1,
  wb = NULL,
  universe = NULL,
  gsl = NULL,
  organism = "UNKNOWN",
  keytype = "UNKNOWN",
  ontology = "UNKNOWN",
  readable = FALSE,
  min_size = 5,
  max_size = 500
)
}
\arguments{
\item{ora_res}{ora analysis result}

\item{pvalueCutoff}{cut off over adjusted p value= 0.25}

\item{pAdjustMethod}{method for p value adjustement}

\item{qvalueCutoff}{cut off over q values}

\item{wb}{input gene list}

\item{universe}{universe of genes}

\item{gsl}{list of gene sets}

\item{organism}{organism}

\item{keytype}{keytype}

\item{ontology}{ontology}

\item{readable}{readable}

\item{min_size}{minimum gene set size}

\item{max_size}{maximum gene set size}
}
\description{
ora2enrich
this function translates the result of an ORA into the DOSE class "enrichResult"
}
