% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora_clusters.R
\name{ora_clusters}
\alias{ora_clusters}
\title{Over Representation Analysis on all clusters of single cell data}
\usage{
ora_clusters(wb, universe, gsl, wb_min = 1, p_adj_method = "fdr", mc_cores = 2)
}
\arguments{
\item{wb}{named list with hits (white balls) vector for each cluster/condition}

\item{universe}{named list with universe genes for each cluster/condition. Will then be used to obtain black balls ("other elements")}

\item{gsl}{named list of sets}

\item{wb_min}{valued used to filter gsl used for testing. If equal to 0, all sets will be tested, otherwise only the ones that 
contains at least `wb_min` `wb` genes will be tested. Useful to limit number of tested sets and avoid testing sets that do not contains `wb` genes}

\item{p_adj_method}{p value adjustment method, see p.adjust.methods}

\item{mc_cores}{number of cores used to parallelize ORA tests}
}
\value{
The function return a named list with a result table for each cluster/condition. In each table the columns are
\itemize{
 \item id: name of the set tested
 \item N: total number of gene tested
 \item wb: number of hit genes
 \item bb: number of other elements present in the test (`N` - `wb`)
 \item bd: number of genes in the set tested
 \item wbd: number of hit genes present in the set tested
 \item exp:
 \item er:
 \item p: p-value
 \item p_adj: corrected p-value calculated by using `p_adj_method`
 \item q_val: q-value calculated using bootstrap method (see qvalue::qvalue for further details)
 \item gene: list of `wbd` genes separated by `;`
}
}
\description{
This function can be used to perform ORA test on multiple clusters/cell-types or conditions
}
