% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_cc.R
\name{pathway_cc}
\alias{pathway_cc}
\title{Caluclate cross-talks between pathways}
\usage{
pathway_cc(pathway_list, gene_network_adj, weight = attr_v$Sp, cluster = 4)
}
\arguments{
\item{pathway_list}{a named list of genes grouped into pathways}

\item{gene_network_adj}{gene network adjacency matrix}

\item{cluster}{numebr of threads to be used}

\item{wight}{an ordered weight vertex vector}
}
\description{
Caluclate cross-talks between pathways
}
