% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_cross_talk.R
\name{pathway_cross_talk}
\alias{pathway_cross_talk}
\title{Caluclate cross-talks between pathways}
\usage{
pathway_cross_talk(
  pathway_list,
  gene_network_adj,
  weight = attr_v$Sp,
  mc_cores_pct = 2,
  mc_cores_perm = 1,
  k = 9
)
}
\arguments{
\item{pathway_list}{a named list of genes grouped into pathways}

\item{gene_network_adj}{gene network adjacency matrix}

\item{mc_cores_pct}{numebr of threads to be used for pathway cross talk calculation}

\item{k}{number of permutation of the adjacency matrix}

\item{wight}{an ordered weight vertex vector}

\item{mc_cores_permutation}{number of thread to be used in permutations}
}
\description{
Caluclate cross-talks between pathways
}
