% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_sim_comm.R
\name{pathway_sim_comm}
\alias{pathway_sim_comm}
\title{Pathway similarity and communities}
\usage{
pathway_sim_comm(
  x,
  gs_list,
  method = c("overlap", "jaccard"),
  coeff = c(0.5, 0.1),
  all_gs = TRUE,
  comm_method = c("auto", "fastgreedy", "multilev"),
  set_sim_df = NULL
)
}
\arguments{
\item{x}{named vector of pathway scores}

\item{gs_list}{gene set list}

\item{method}{'overlap' or 'jaccard'}

\item{coeff}{threshold for the similarity score between two gene sets}

\item{all_gs}{TRUE/FALSE}

\item{comm_method}{one among 'auto', 'fastgreedy' and 'multilev'}

\item{set_sim_df}{optional data frame with three columns 'set1', 'set2' and 'sim'}
}
\value{
list of two data frames containing pathway network and vertex similarity
}
\description{
Pathway similarity and communities
}
