% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_similarity.R
\name{pathway_similarity}
\alias{pathway_similarity}
\title{Pathway similarity}
\usage{
pathway_similarity(pathway_list, gene_network_adj, sim_method = "overlap")
}
\arguments{
\item{pathway_list}{list of gene sets}

\item{gene_network_adj}{adjacency matrix of the gene network under analysis}

\item{sim_method}{method to define the similarity between pathways}
}
\description{
Pathway similarity
}
