% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_link.R
\name{perm_link}
\alias{perm_link}
\title{Function to caluclate permutation of the number of links between two gene sets}
\usage{
perm_link(r, c, gene_network_adj, core, k, hash = T)
}
\arguments{
\item{r}{length of the first geneset (rows)}

\item{c}{length of the second geneset (columns)}

\item{gene_network_adj}{adjacency matrix of the gene network used}

\item{core}{threads to be used for permutation}

\item{k}{number of permutation}

\item{hash}{logical, indicating if random sampling should be done by using hashmap (see `sample.int()`)}
}
\value{
The function a list k matrices, each one composed by the number of links in the permutated verion of the 
two genesets
}
\description{
`perm_link` is used to obtain permutatied number of links between two genesets of 
a defined number of gene
}
\details{
The function generates k permuted verion of two geneset of defined length by random sampling the 
gene network adjacency matrix. Then identify the subset of the adjacency matrix for each k geneset and 
calculates the number of links.
}
