% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_communication_network.R
\name{plot_communication_network}
\alias{plot_communication_network}
\title{Cell-cell communication network plot}
\usage{
plot_communication_network(
  cell_emb,
  cl_res,
  cl_cell,
  palette,
  e_scale = 9,
  label,
  save_file,
  point_size = 0.5,
  point_alpha = 0.5,
  edge_alpha = 0.8,
  edge_color = "gray45",
  node_size = 5,
  text_size = 2,
  text_color = "black",
  heigh = 100,
  width = 100,
  res = 300,
  unit = "mm"
)
}
\arguments{
\item{cell_emb}{cell embedding (UMAP or tSNE) matrix}

\item{cl_res}{cell-cell communication results (from `gs_cross_talk()`)}

\item{cl_cell}{two column data-frame, the first should be the cell names, the second the annotation (cluster or cell types)}

\item{palette}{named vector of the color associated to each cluster or cell types, and it will be used to set both 
cell points and communication network vertices color. If NULL ignored, ggplot automatic palette used}

\item{e_scale}{value to scale the edge weights for plotting, change to adjust maximum link widths}

\item{label}{logical, if cluster/cell type name should be plotted on the vertex names}

\item{save_file}{saving plot name. If NULL ignored}

\item{point_size}{= 0.5 size of the cell points to be plotted.}

\item{point_alpha}{= 0.5 transparency of the color fill of the cell points}

\item{edge_alpha}{= 0.8 transparency of the edge color}

\item{edge_color}{= edge color}

\item{node_size}{= 5 size of the vertices of the cell-cell communication network overlapped to the cell points}

\item{text_size}{size of the node labels plotted (if `label = T`)}

\item{text_color}{= text color (if `label = T`)}

\item{heigh, width, res, unit}{params used to save plot (in jpeg format)}
}
\description{
This function is used to plot the results of the cell-cell communication analysis
 overlapped to the single-cell plot
}
\details{
This functions uses the cell-embeddings produced by the single-cell analysis
 and the cell-cell communication analysis results, obtained with `gs_cross_talk()`, to plot
 the communication network overlapped to the cell plot. This functions do not filter the 
 `gs_cross_talk()` results to plot only the significant ones.
 Cells and communication network vertices are colored according to the annotation, provided with `cl_cell`
}
