% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functional_relevance.R
\name{plot_functional_relevance}
\alias{plot_functional_relevance}
\title{Function to plot functional relevance results}
\usage{
plot_functional_relevance(
  fr,
  method = "count",
  file_name = NULL,
  plot_names = T,
  pal = NULL,
  jitter_width = 0.07,
  jitter_height = 0.07,
  width = 200,
  height = 200,
  res = 300,
  units = "mm"
)
}
\arguments{
\item{fr}{the data.frame obtained as `gene_functional_relevance()` result}

\item{method}{= c("count", "relative") If "count" then the functional and interactor diversity and functional relevance are used, otherwise the relative ones}

\item{file_name}{file name of the plot, if `NULL` the plot will be returned by the function}

\item{plot_names}{logical, or a vector with the same length and order of the genes in `fr`. If `TRUE` all genes name are plotted,
if `FALSE` no names are plotted, otherwise the passed vector is used to label genes and all of them are plotted}

\item{pal}{vector of three color to be used to create the gradient to color the genes according to (relative) functional relevance score. If `NULL`
the Brewer "Spectral" palette is used}

\item{jitter_width, jitter_height}{values to control point jitter in the plot. See ggplot2::geom_jitter() for details}

\item{width, height, res, units}{values used to save the plot in jpeg format, see grDevices::jpeg() for details}

\item{ct_null}{required only if `method = "normalized"`. This should be a cross-talk result table coming from a general model}
}
\value{
If `file_name` The function returns the ggplot object
}
\description{
Graphical function to plot (relative) functional relevance results.
}
\details{
The function plots (relative) functional relevance results obtained with `gene_functional_relevance()` function. Each gene is plotted by interactor
 and functional diversity, and is colored by functional relevance score
}
