% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea_heatmap.R
\name{plot_gsea_heatmap}
\alias{plot_gsea_heatmap}
\title{Plot heatmap of GSEA results for multiple runs}
\usage{
plot_gsea_heatmap(
  gsea_res = NULL,
  nes_sign = TRUE,
  a = 0.25,
  p.stat = "FDRq",
  na_col = "khaki",
  max_gs = 50,
  min.p = 1e-04,
  ...
)
}
\arguments{
\item{gsea_res}{result of function gsea()}

\item{nes_sign}{whether to print nes sign}

\item{a}{alpha over FDR}

\item{na_col}{color for FDR > a}

\item{max_gs}{maximum number of gene sets that will be plotted}

\item{...}{further arguments to ComplexHeatmap::Heatmap()}
}
\description{
Plot heatmap of GSEA results for multiple runs
}
