% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ora_comparison.R
\name{plot_ora_comparison}
\alias{plot_ora_comparison}
\title{Plot a comparion between two ORAs}
\usage{
plot_ora_comparison(
  ora_res_list = NULL,
  p_sig = 0.001,
  p_max = 0.1,
  dir_out = "./",
  mar = c(2.5, 20, 1, 1),
  mgp = c(1.2, 0.3, 0),
  cex.axis = 0.6,
  use_nominal_p = TRUE,
  col_pal = NULL
)
}
\arguments{
\item{ora_res_list}{list of results of ora_pipeline}

\item{p_sig}{thresold for significant pathways}

\item{p_max}{maximum p-value allowed for marginally significant pathways}

\item{dir_out}{output directory}

\item{mar}{mar parameter}

\item{mgp}{mgp parameter}

\item{cex.axis}{cex.axis}

\item{use_nominal_p}{whether to use or not nominal p-values}

\item{col_pal}{color palette function}
}
\value{
data.frame with merged results only for pathways that satisfy p_sig in at least one condition
}
\description{
Plot a comparion between two ORAs
}
