% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_gs_list.R
\name{preparing_gs_list}
\alias{preparing_gs_list}
\title{Build list for cross-talk calculation from database list}
\usage{
preparing_gs_list(gs_names, gs_genes, weights, min_size = 1, max_size = 500)
}
\arguments{
\item{gs_names}{names of the gene-sets, to be used to build the gene-set list. Should have the same length of `gs_gene`}

\item{gs_genes}{names of the genes in the gene-sets `gs_genes`}

\item{weights}{named vector with a weight for each gene in `gs_gene`. If `NULL` the function assigns `1` to each gene}
}
\value{
The output is a gene-set list with a vector for each gene-set, composed by weights named after the genes
}
\description{
The function build a gene-set list compatible with `gs_cross-talk()` for cross-talk calculation starting from a provided gene set database
 and genes (with weights) of interest.
}
\details{
`preparing_gs_list()` uses a gene-set database data to build the list needed by `gs_cross_talk()` for cross-talk calculation.
}
