% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_msigdb_list.R
\name{preparing_msigdb_list}
\alias{preparing_msigdb_list}
\title{Build list for cross-talk calculation from MSigDB database list}
\usage{
preparing_msigdb_list(
  species,
  category = NULL,
  subcategory = NULL,
  type = "gene_symbol",
  genes,
  weights,
  min_size = 1,
  max_size = 500
)
}
\arguments{
\item{species}{Species name}

\item{category}{MSigDB category name, see `msigdbr_collections()`}

\item{subcategory}{MsigDB subcategory name, see `msigdbr_collections()`}

\item{type}{Gene name of interest, can be `gene_symbol` or `entrez_gene`}

\item{genes}{vector with gene gene name of interest}

\item{weights}{vector with `genes` weights in the same order. If `NULL` the function assigns `1` to each gene}

\item{min_size, max_size}{filtering the gene sets to maintain only the ones that have at least `min_size` or less than `max_size` number of genes}
}
\value{
The output is a gene-set list with a vector for each gene-set, composed by weights named after the genes
}
\description{
The function build a gene set list compatible with `gs_cross-talk()` for cross-talk calculation using a desired MSigDB (using msigdbr package)
}
\details{
`preparing_msigdb_list()` uses a MSigDB database to build the list needed by `gs_cross_talk()` for cross-talk calculation. In details, it calls
 `pathway_data()` function to download the database by using msigdbr package to then build the gene-set list
}
