% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_correlation.R
\name{result_correlation}
\alias{result_correlation}
\title{Function to compare results by using Spearman correlation}
\usage{
result_correlation(res_list)
}
\arguments{
\item{res_list}{a named list of data.frame, each composed by three columns: first and second elements of the cross-talk, 
score column.}
}
\value{
The function returns a correlation matrix named after the names of `res_list`
}
\description{
The function calculates the Spearman correlation between each pair of the provided results
}
\details{
The function takes as an input a list of elements, each  composed by a data.frame with three columns:
name of the first element (pathway or cell cluster), name of the second and the score that will be used for the comparison.
The function calculates the correlation score between each possible pairs and returns as an output a correlation matrix
where each rows and columns are named after the names of the elements of the list.
}
