% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_network.R
\name{result_network}
\alias{result_network}
\title{Function to compare networks of PCT or CCC using Jaccard index on the edges and/or Structural similarity}
\usage{
result_network(res_list, type = c("jaccard", "structural"))
}
\arguments{
\item{res_list}{a named list of networks}

\item{type}{which measure to calcualte, can be either "jaccard", "strcutural" o both}
}
\value{
The function returns a matrix for each mesure calcualted. The matrix has rows and column numbers equal to the
number of elements in `res_list` and named after the names of the elements in it
}
\description{
This functionis used to calculate Jaccard index and/or Structural similarity to compare multiple PCTN or CCCN
}
\details{
This function takes as an input a list of networks. Then it calcuates Jaccard index on the edges and/or 
Structural similarityes between all pairs
}
