% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_integrated_cross_talk.R
\name{single_integrated_cross_talk}
\alias{single_integrated_cross_talk}
\title{Integrated cross talk on specific cluster analysis}
\usage{
single_integrated_cross_talk(
  S_list,
  ref,
  A,
  k = 9,
  bin_type = "number",
  cut_par = 9,
  perm_link = "degree",
  perm_weights = "simple",
  mc_cores_perm = 1,
  mc_cores_ct = 1
)
}
\arguments{
\item{S_list}{a named list of genes grouped into gene sets, as obtained from `build_S_list()`, plus the reference gene set list}

\item{ref}{name of the gene set in `S_list` that contains all genes of cluster i involved in the communication with cluster j}

\item{A}{adjacency matrix of the whole gene network considered (can be a sparseMatrix)}

\item{k}{number of permutations}

\item{bin_type}{can be either "number" (suggested), "interval" or "width". See `ggplot2::cut_interval()` for details.}

\item{cut_par}{number of bin to cut. If set to `NULL`, the function will search for the best cut in 2:15}

\item{perm_link}{= c("degree", "simple") If the permutation of the link should be degree-conservative ("degree"), or random ("simple")}

\item{perm_weights}{= c("degree", "simple") If the permutation of the weights should be degree-conservative ("degree"), or random ("simple")}

\item{mc_cores_perm}{number of thread to be used in permutations}

\item{mc_cores_ct}{number of threads to be used for cross talk calculation}
}
\value{
The function returns a data.frame with the results of ct calculation:
\itemize{
   \item S1_name,S2_name: names of gene-set pairs considered
   \item c: cross-talk score
   \item S1_size,S2_size: number of genes in S1 and S2
   \item S1_s2_size,S2_s1_size: number of S1 genes interacting with S2, and vice versa
   \item dL: number of links between S1 and S2
   \item L: number of possible links between S1 and S2
   \item r_c: cross-talk saturation, calculated as `dL/L`
   \item u1,u2: sum of the gene weights in S1 and S2, respectively
   \item S1,S2: list of interacting genes in S1 and S2, respectively
   \item s: cross-talk summary score
   \item pA,pU: p-values of the number of links (pA) and weights (pU)
   \item p: combined p-value
 }
}
\description{
The function calculates cross-talk between genes involved in communication with a specific cluster/cell type and cluster gene sets
}
\details{
The function takes as inputs the adjacency matrix of the biological network (`A`), the gene-set list `S_list()` and the name of the gene set that
 contains the communication genes (`ref`). For each gene-set pair that shows at least a link between them, the function calculates the cross-talk 
 and the statistical evaluation, as described in the paper, using `k` permutations.
}
