% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM_PCT.R
\name{TM_PCT}
\alias{TM_PCT}
\title{Caluclate cross-talks between pathways}
\usage{
TM_PCT(
  pathway_list,
  gene_network_adj,
  membership = attr_v$comm_id,
  mc_cores_pct = 2,
  mc_cores_perm = 2,
  mc_cores_tm = 2,
  weight = attr_v$Sp,
  k = 9
)
}
\arguments{
\item{pathway_list}{a named list of genes grouped into pathways}

\item{gene_network_adj}{gene network adjacency matrix}

\item{membership}{an ordered vector of vertex membership to a topological community}

\item{mc_cores_pct}{numebr of threads to be used to calculate cross talk between different communities}

\item{mc_cores_perm}{numebr of threads to be used to calcualte permutation}

\item{mc_cores_tm}{number of threads to be used to calcualte TM-PCT on different communities combination}

\item{k}{number of permutation of the adjacency matrix}

\item{wight}{an ordered weight vertex vector}
}
\description{
Caluclate cross-talks between pathways
}
