% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uOra2enrich.R
\name{uOra2enrich}
\alias{uOra2enrich}
\title{ulisse2enrich}
\usage{
uOra2enrich(
  ulisse_res,
  pvalueCutoff = 0.25,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.25,
  gene = NULL,
  universe = NULL,
  geneSets = NULL,
  organism = "UNKNOWN",
  keytype = "UNKNOWN",
  ontology = "UNKNOWN",
  readable = FALSE
)
}
\arguments{
\item{ulisse_res}{ora analysis result}

\item{pvalueCutoff}{cut off over adjusted p value= 0.25}

\item{pAdjustMethod}{method for p value adjustement}

\item{qvalueCutoff}{cut off over q values}

\item{gene}{input gene list}

\item{universe}{universe of genes}

\item{geneSets}{list of gene sets}

\item{organism}{organism}

\item{keytype}{keytype}

\item{ontology}{ontology}

\item{readable}{readable}
}
\description{
ulisse2enrich
}
