% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p.R
\name{calc_p}
\alias{calc_p}
\title{Estimation of p values}
\usage{
calc_p(X)
}
\arguments{
\item{X}{list of matrices, where the first is the one obtained with real data and the others are obtained through permutations. 
The matrices should be 1 column matrices with the same order.}
}
\description{
This function is used to calculate p-value from a list of matrices
}
