% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_set_similarity.R
\name{calc_set_similarity}
\alias{calc_set_similarity}
\title{The function calculates similarities between two sets.}
\usage{
calc_set_similarity(x, y, method = c("jaccard", "overlap", "average"))
}
\arguments{
\item{x, y}{two sets provided as vectors}

\item{method}{one among "jaccard", "overlap" and "average"}
}
\value{
similarity
}
\description{
The function takes as input two vertices and calculates their similarity by using three different methods:
\itemize{
\item Jaccard = number of common elements between x and y, divided by the number of unique elements present in c(x, y)
\item Overlap = number of common elements between x and y. divided by the minimum length of unique elements present in either x or y
\item Average = the mean between Jaccard and Overlap measure
}
}
\details{
`calc_set_similarity` calculates similarities between two sets by using `jaccard`, `overlap`, `average` methods.
}
