% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_map_with_plot.R
\name{enrichment_map}
\alias{enrichment_map}
\title{Enrichment map}
\usage{
enrichment_map(
  x,
  gs_list,
  method = c("overlap", "jaccard"),
  coeff = NULL,
  all_gs = TRUE,
  weight.within = 4,
  weight.between = 1,
  file.prefix = "en_map",
  comm_method = "fastgreedy",
  vertex.label.cex = 1,
  vertex.label.dist = 0.5,
  vertex.label.degree = pi/2,
  vertex.label.font = 2,
  gs_list_size = NULL,
  set_sim_df = NULL,
  plot_flag = TRUE,
  vertex.color = NULL,
  img.width = 200,
  img.height = 200,
  img.res = 300,
  min_comm_size = 2,
  vertex.size.min = 2,
  vertex.size.max = 5,
  edge.wd.min = 1,
  edge.wd.max = 4,
  mark.groups = T,
  vertex.shape = "circle",
  vertex.pie = NULL,
  vertex.color.pal = NULL,
  n_width = 3,
  n_size = 3,
  L = NULL,
  score.decreasing = TRUE,
  n_name = 1,
  legend.cex = 0.7,
  ...
)
}
\arguments{
\item{x}{named vector of pathway scores}

\item{gs_list}{gene set list}

\item{method}{'overlap' or 'jaccard'}

\item{coeff}{threshold for the similarity score between two gene sets}

\item{all_gs}{TRUE/FALSE}

\item{weight.within}{value to weight the attraction between two vertices of the same community}

\item{weight.between}{value to weight the attraction between two vertices of two disntict communities}

\item{comm_method}{one among 'auto', 'fastgreedy' and 'multilev'}

\item{vertex.label.cex}{vertex label size}

\item{vertex.label.dist}{vertex label distance from veritex}

\item{vertex.label.degree}{vertex label orientation}

\item{vertex.label.font}{vertex label font}

\item{gs_list_size}{gene set list size}

\item{set_sim_df}{optional data frame with three columns 'set1', 'set2' and 'sim'}

\item{plot_flag}{TRUE/FALSE}

\item{vertex.color}{vertex color}

\item{img.width}{image width}

\item{img.height}{image height}

\item{img.res}{image resolution}

\item{min_comm_size}{min community size}

\item{vertex.size.min}{min vertex size}

\item{vertex.size.max}{max vertex size}

\item{edge.wd.min}{min edge widht}

\item{edge.wd.max}{max edge width}

\item{mark.groups}{TRUE/FALSE}

\item{vertex.shape}{vertex shape}

\item{vertex.pie}{vertex pie}

\item{...}{additional parameters of plot.igraph}

\item{file}{output file}
}
\value{
list of two data frames containing pathway network and vertex similarity
}
\description{
Enrichment map
}
