% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_funct_relevance.R
\name{gene_funct_relevance}
\alias{gene_funct_relevance}
\title{Function to calculate gene functional relevance}
\usage{
gene_funct_relevance(pct, adj, to_plot = T, file_name = NULL, plot_names = T)
}
\arguments{
\item{pct}{the data.frame obtained as a result of `pathway_corss_talk()`}

\item{adj}{the adjacency matrix used as an input in `pathway_corss_talk()`}

\item{to_plot}{logical, indicating whether to save the plot of functional relevance}

\item{file_name}{file name of the plot, if NULL the plot will be saved in the working 
directory as "gene_functional_relevance.jpeg"}

\item{plot_names}{logical, indicating whether to print gene names in the functional relevance plot}
}
\value{
The function returns a data.frame with:
\itemize{
\item gene: the gene analysed
\item nPTW: functional diversity
\item nInteractors: interactor diversity
\item pathway: list of names of the pathway counted in `nPTW` separated by `;`
\item interactors: list of names of the genes counted in `nInteractors` spearated by `;`
}
}
\description{
gene_funct_relevance() calculates functional diversity and interactor diversity for the gene 
that participate in significative pathway cross-talks
}
\details{
gene_funct_relevance() takes as an input the data.frame obtained from `pathway_cross_talk()` and the 
adjacency matrix used as an input of `pathway_cross_talk()`. The function then extract the genes partecipating in
significative cross-talks and calculate for each of them the functional diversity and the interactor diversity. 
Considering a gene, the functional diversity is the number of pathway with which the gene has a link and participate 
in the formation of a significative cross-talk; the interactor diversity is the number of gene with which the gene 
has a link that contibute in the formation of a significative cross-talk.
}
\examples{
ptw_list <- list(ptwA = c("A", "B","C"), ptwB = c("D", "E", "F"), ptwC = c("A", "B", "E"))
adj <- matrix(data = sample(c(0,1), 6*6, replace = T), nrow = 6, 
ncol = 6, dimnames = list(LETTERS[1:6], LETTERS[1:6]))
wgt <- rep(1, 6)
pct <- pathway_cross_talk(pathway_list = ptw_list, gene_network_adj = adj, weight = wgt, 
  mc_cores_pct = 1, mc_cores_perm = 1, k = 9)
funct_rel <- gene_funct_relevance(pct, adj)
}
