% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_cc.R
\name{pathway_cc}
\alias{pathway_cc}
\title{Function to caluclate sub-components pathway score}
\usage{
pathway_cc(pathway_list, gene_network_adj, genes, weight, mc_cores_cc = 2)
}
\arguments{
\item{pathway_list}{a named list of genes grouped into pathways}

\item{gene_network_adj}{adjacency matrix of the whole gene network considered}

\item{genes}{vector with the gene of interest to be used for pathway cc calculation}

\item{weight}{weights of the genes in pathway list. If not provided, the function assign to each gene
a weight of 1}

\item{mc_cores_cc}{numebr of threads to be used for cc calculation}
}
\value{
The function returns a list of two object:
\enumerate{
 \item membership: results of [igraph::components()] for each pathway considered
 \item {pathway_cc}: a data frame with
\itemize{
 \item pathway: name of the pathway where the components scores are calcualted
 \item ID: id of the component of the pathway
 \item score: component score
 \item n_gene: number of genes in the component
 \item n_link: number of links in the component
 \item gene: names of the genes in the component
}
}
}
\description{
Calculates the score of each pathway connected components
}
\details{
By coupling the pathway data to the adjacency matrix, the function identifies the pathway components as
 connected components (see [igraph::components()] for details). For each component, the function calcualtes the score 
 as described in the article.
}
\examples{
\dontrun{
ptw_list <- list(ptwA = c("A", "B","C"), ptwB = c("D", "E", "F"), ptwC = c("A", "B", "E"))
adj <- matrix(data = sample(c(0,1), 6*6, replace = TRUE), nrow = 6, 
ncol = 6, dimnames = list(LETTERS[1:6], LETTERS[1:6]))
wgt <- rep(1, 6)
p_cc <- pathway_cc(pathway_list = ptw_list, 
gene_network_adj = adj, weight = wgt, mc_cores_cc = 1)
}
}
