% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_cross_talk.R
\name{pathway_cross_talk}
\alias{pathway_cross_talk}
\title{Pathway Cross Talk}
\usage{
pathway_cross_talk(pathway_list, gene_network_adj)
}
\arguments{
\item{pathway_list}{list of gene sets}

\item{gene_network_adj}{adjacency matrix of the gene network under analysis}
}
\description{
Pathway Cross Talk
}
