% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_cross_talk.R
\name{pathway_cross_talk}
\alias{pathway_cross_talk}
\title{Function to caluclate cross-talks between pathways}
\usage{
pathway_cross_talk(
  pathway_list,
  gene_network_adj,
  genes,
  weight = NULL,
  mc_cores_pct = 2,
  mc_cores_perm = 1,
  k = 9
)
}
\arguments{
\item{pathway_list}{a named list of genes grouped into pathways}

\item{gene_network_adj}{adjacency matrix of the whole gene network considered (can be a sparseMatrix)}

\item{genes}{target gene to be used for PCT calculation}

\item{weight}{an vector of weights for each gene the pathway list. If not provided, the function assigns to each gene
a weight of 1}

\item{mc_cores_pct}{numebr of threads to be used for pathway cross talk calculation}

\item{mc_cores_perm}{number of thread to be used in permutations}

\item{k}{number of permutation}
}
\value{
The function returns a table with:
\itemize{
 \item pathway_1, pathway_2: the pathway pair considered
 \item pct: the CT value
 \item ngenes_pathway_1, ngenes_pathway_2: number of genes involved in `pathway_1` and `pathway_2`, respectively
 \item nlink: number of links between the genes in `pathway_1` and `pathway_2`
 \item: weight_pathway1, weight_pathway2: cumulative weights of the genes involved in the 
 cross-talk between pathway 1 and 2
 \item p_value_link: p-value score calcualted by using the permutation approach
 \item FDR_link: empirical FDR
 \item p_adj_BH: `p-value_link` adjusted using BH method
 \item gene_pathway1, gene_pathway2: gene involevd in the CT in `pathway_1` and `pathway_2`, respectively
}
}
\description{
Calculates cross-talk between all pathway pairs that show a link between their exclusive genes
}
\details{
The function defines a subset of the `gene_network_adj` for each pathway pair, 
 with the exclusive genes of a pathway in the rows and the exclusive genes of the other in the columns. 
 Only the pathway pairs that show a link in these subset will be considered. Then, the CT is calcluated, 
 as described in the paper. This approach is applied to the original adjacency matrix and to all its permuted versions.
 The permutations are obtained by random sampling genes from the adjacency matrix for each pathway.
 The p-value and FDR are calculated on the number of links, without considering the weights of the genes.
 We also provide a corrected p-value using BH method.
}
\examples{
 
 ptw_list <- list(ptwA = c("A", "B","C"), ptwB = c("D", "E", "F"), ptwC = c("A", "B", "E"))
 adj <- matrix(data = sample(c(0,1), 20*20, replace = TRUE), nrow = 6, 
 ncol = 6, dimnames = list(LETTERS[1:20], LETTERS[1:20]))
 pct <- pathway_cross_talk(pathway_list = ptw_list, gene_network_adj = adj, weight = NULL, 
 genes = LETTERS[1:6], mc_cores_pct = 1, mc_cores_perm = 1, k = 9)
}
