% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_data.R
\name{pathway_data}
\alias{pathway_data}
\title{Returns filtered msigdb geneset in a format compatible with cross-talks calculations}
\usage{
pathway_data(
  species,
  category = NULL,
  subcategory = NULL,
  type = "gene_symbol",
  genes,
  min_size = 1,
  max_size = 500
)
}
\arguments{
\item{species}{Species name}

\item{category}{MSigDB category name, see `msigdbr_collections()`}

\item{subcategory}{MsigDB subcategory name, see `msigdbr_collections()`}

\item{type}{Gene name of interest, can be `gene_symbol` or `entrez_gene`}

\item{genes}{vector with gene gene name of interest}

\item{min_size, max_size}{filtering the gene sets to maintain only the ones that have at least `min_size` or less than `max_size` number of genes}
}
\value{
The output is a list of two output:
\enumerate{
\item msigdb_output: a data.frame with the data downloaded by MSigDB data
\item path_list: a list of pathway. Each list is composed by a vector with the names of the genes of interest 
that are part of that pathway
}
}
\description{
Download MsidgDB data, filter them and prepare them in the format needed by cross-talks calculation
}
\details{
The function downloads the pathway data for the `species` of interest. The selection of the database to 
download is made by using the parameter `category` and `subcategory`. The data downloaded are then filtered to
maintain only the `genes` of interest and to remove pathway with a number of genes lower of `min_size` and higher of
`max_size`
}
