% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_CT.R
\name{plot_network_CT}
\alias{plot_network_CT}
\title{Function for plotting results from PCT or CCC analysis as a network}
\usage{
plot_network_CT(
  ct,
  all = FALSE,
  p_val,
  FDR,
  ct_val,
  community,
  pal_community = NULL,
  edge_col_by = "pct",
  edge_pal = NULL,
  edge_value = NULL,
  edge_breaks = 1,
  file_out = NULL,
  width = 200,
  height = 200,
  res = 300,
  units = "mm",
  layout = NULL,
  legend_edge_col = T,
  legend_filt = T,
  legend_community = T,
  legend_edge_breaks = T,
  x_filt = -1.5,
  y_filt = 0,
  x_comm = -1.5,
  y_comm = -0.5,
  x_edge = -1.5,
  y_edge = -1,
  xl = -1.45,
  yb = 0.2,
  xr = -1.3,
  yt = 0.5,
  ...
)
}
\arguments{
\item{ct}{output of `pathway_cross_talk()` or `cluster_communication()` function}

\item{all}{logical, if the function have to plot all the edges or only the significative ones. If FALSE, if the user
provides filtering parameters, the function plot the CT passing the threshold as a solid line, 
the others as a dashed line}

\item{p_val, FDR, ct_val}{filtering values. If one of these is set to NULL the function ignores it}

\item{community}{logical or an object resulting from igraph community calculation. If TRUE, the function calculates
the communities by using fastgreedy algorithm, if FALSE the function does not consider communities. Otherwise, if a
community object is provided, the function use it and color the vertices accordingly}

\item{pal_community}{color palette to be used to color the vertices according to the communities. If not provided,
the function uses `rainbow()` palette}

\item{edge_col_by}{name of the column to be used to color the edges. If NULL the edges will be not colored}

\item{edge_pal}{color to be used to create the gradient to color the edges. If not provided, the function uses "blue" 
to "red"}

\item{edge_value}{vector of the two values to be used to create the gradient to color the edges. If not provided, 
the function uses minimum and maximum}

\item{edge_breaks}{number of breaks to be used to group values in `edge_col_by` and plot the edges with different 
width. If set to 1, all the edges will have the same width}

\item{file_out}{name of the jpeg file produced. If not provided, the plot will be automatically named "ct_network.jpeg}

\item{width, height, res, units}{graphical value of `jpeg()` function}

\item{layout}{layout to be used in igraph plotting. If not provided, the function calculates it}

\item{legend_edge_col}{logical, if to plot the legend of the color of the edges}

\item{legend_filt}{logical, if to plot legend of the different lines used when `all = T`}

\item{legend_community}{logical, if to plot the legend of the community color used}

\item{legend_edge_breaks}{logical, if to plot the legend of the different interval used for edge width plotting}

\item{x_filt, y_filt, x_comm, y_comm, x_edge, y_edge}{x and y coordinates for `legend_filt`, `legend_community`, 
`legend_edge` positioning in the plot. See `legend()` function for more details}

\item{xl, yb, xr, yt}{coordinates for the positioning of the gradient legend for `legend_edge_col`. See `color.legend()` 
and `rect()` functions for more details}

\item{...}{further graphical parameters to be passed to `igraph.plot()` function}
}
\value{
the function produce the plot saved with the name passed to `file_out` and also returns the layout,
the igraph object used for plotting and the communities (if calculated)
}
\description{
The function elaborate the result of `pathway_cross_talk()` or `cluster_communication()` function to obtain a graphical 
representation of the cross-talk network
}
\details{
The functions uses `pathway_cross_talk()` or `cluster_communication()` output to build a CT network. Then, the network is plotted with
the edges colored by a value of interest. The user may decide if to plot all the edges, color the vertices by 
their community, scale the dimension of the edges over a variable of interest.
}
