% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ora_heatmap.R
\name{plot_ora_heatmap}
\alias{plot_ora_heatmap}
\title{Plot heatmap of ORA results for multiple runs}
\usage{
plot_ora_heatmap(
  ora_res = NULL,
  p.stat = "p_adj",
  a = 0.25,
  na_col = "khaki",
  max_gs = 50,
  ...
)
}
\arguments{
\item{ora_res}{result of function gsea()}

\item{p.stat}{p-value that will be plotted}

\item{a}{alpha over FDR}

\item{na_col}{color for FDR > a}

\item{max_gs}{maximum number of gene sets that will be plotted}

\item{...}{further arguments to ComplexHeatmap::Heatmap()}
}
\description{
Plot heatmap of ORA results for multiple runs
}
