% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM_PCT.R
\name{TM_PCT}
\alias{TM_PCT}
\title{Caluclate cross-talks between pathways}
\usage{
TM_PCT(
  pathway_list,
  gene_network_adj,
  membership = attr_v$comm_id,
  weight = attr_v$Sp,
  cluster = 4
)
}
\arguments{
\item{pathway_list}{a named list of genes grouped into pathways}

\item{gene_network_adj}{gene network adjacency matrix}

\item{membership}{an ordered vector of vertex membership to a topological community}

\item{cluster}{numebr of threads to be used}

\item{wight}{an ordered weight vertex vector}
}
\description{
Caluclate cross-talks between pathways
}
